refactor: exact imports for attribute types
