+++ packages/consumption/src/modules/attributes/events/AttributeCreatedEvent.ts
@@ -1,5 +1,5 @@
 import { TransportDataEvent } from "@nmshd/transport";
-import { LocalAttribute } from "../local/attributeTypes";
+import { LocalAttribute } from "../local/attributeTypes/LocalAttribute";
 
 export class AttributeCreatedEvent extends TransportDataEvent<LocalAttribute> {
     public static readonly namespace = "consumption.attributeCreated";
