+++ packages/consumption/src/modules/attributes/events/AttributeWasViewedAtChangedEvent.ts
@@ -1,5 +1,5 @@
 import { TransportDataEvent } from "@nmshd/transport";
-import { LocalAttribute } from "../local/attributeTypes";
+import { LocalAttribute } from "../local/attributeTypes/LocalAttribute";
 
 export class AttributeWasViewedAtChangedEvent extends TransportDataEvent<LocalAttribute> {
     public static readonly namespace = "consumption.attributeWasViewedAtChanged";
