+++ packages/consumption/src/modules/attributes/index.ts
@@ -1,6 +1,5 @@
 export * from "./AttributesController";
 export * from "./events";
 export * from "./local/AttributeTagCollection";
-export * from "./local/attributeTypes";
 export * from "./local/sharingDetails";
 export * from "./local/successorParams";
