--- packages/consumption/src/modules/attributes/local/attributeTypes/index.ts
@@ -1,8 +1,0 @@
-export * from "./AttributeWithPeerSharingDetails";
-export * from "./ForwardableAttribute";
-export * from "./LocalAttribute";
-export * from "./OwnIdentityAttribute";
-export * from "./OwnRelationshipAttribute";
-export * from "./PeerIdentityAttribute";
-export * from "./PeerRelationshipAttribute";
-export * from "./ThirdPartyRelationshipAttribute";
