+++ packages/consumption/src/modules/requests/itemProcessors/proposeAttribute/ProposeAttributeRequestItemProcessor.ts
@@ -14,15 +14,10 @@ import { CoreAddress } from "@nmshd/core-types";
 import { TransportCoreErrors } from "@nmshd/transport";
 import { nameof } from "ts-simple-nameof";
 import { ConsumptionCoreErrors } from "../../../../consumption/ConsumptionCoreErrors";
-import {
-    AttributeSucceededEvent,
-    OwnIdentityAttribute,
-    OwnIdentityAttributeSuccessorParams,
-    PeerIdentityAttribute,
-    PeerIdentityAttributeSuccessorParams,
-    ReceivedAttributeDeletionStatus
-} from "../../../attributes";
-import { LocalAttribute } from "../../../attributes/local/attributeTypes";
+import { AttributeSucceededEvent, OwnIdentityAttributeSuccessorParams, PeerIdentityAttributeSuccessorParams, ReceivedAttributeDeletionStatus } from "../../../attributes";
+import { LocalAttribute } from "../../../attributes/local/attributeTypes/LocalAttribute";
+import { OwnIdentityAttribute } from "../../../attributes/local/attributeTypes/OwnIdentityAttribute";
+import { PeerIdentityAttribute } from "../../../attributes/local/attributeTypes/PeerIdentityAttribute";
 import { ValidationResult } from "../../../common/ValidationResult";
 import { GenericRequestItemProcessor } from "../GenericRequestItemProcessor";
 import { LocalRequestInfo } from "../IRequestItemProcessor";
