+++ packages/consumption/src/modules/requests/outgoing/OutgoingRequestsController.ts
@@ -8,7 +8,10 @@ import { ConsumptionControllerName } from "../../../consumption/ConsumptionContr
 import { ConsumptionCoreErrors } from "../../../consumption/ConsumptionCoreErrors";
 import { ConsumptionError } from "../../../consumption/ConsumptionError";
 import { ConsumptionIds } from "../../../consumption/ConsumptionIds";
-import { EmittedAttributeDeletionInfo, EmittedAttributeDeletionStatus, OwnIdentityAttribute, OwnRelationshipAttribute, PeerRelationshipAttribute } from "../../attributes";
+import { EmittedAttributeDeletionInfo, EmittedAttributeDeletionStatus } from "../../attributes";
+import { OwnIdentityAttribute } from "../../attributes/local/attributeTypes/OwnIdentityAttribute";
+import { OwnRelationshipAttribute } from "../../attributes/local/attributeTypes/OwnRelationshipAttribute";
+import { PeerRelationshipAttribute } from "../../attributes/local/attributeTypes/PeerRelationshipAttribute";
 import { ValidationResult } from "../../common/ValidationResult";
 import { OutgoingRequestCreatedAndCompletedEvent, OutgoingRequestCreatedEvent, OutgoingRequestStatusChangedEvent } from "../events";
 import { RequestItemProcessorRegistry } from "../itemProcessors/RequestItemProcessorRegistry";
