+++ packages/runtime/src/dataViews/DataViewExpander.ts
@@ -1,14 +1,11 @@
 import { Serializable, SerializableBase } from "@js-soft/ts-serval";
-import {
-    ConsumptionController,
-    ForwardableAttribute,
-    LocalRequestStatus,
-    OwnIdentityAttribute,
-    OwnRelationshipAttribute,
-    PeerIdentityAttribute,
-    PeerRelationshipAttribute,
-    ThirdPartyRelationshipAttribute
-} from "@nmshd/consumption";
+import { ConsumptionController, LocalRequestStatus } from "@nmshd/consumption";
+import { ForwardableAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/ForwardableAttribute";
+import { OwnIdentityAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/OwnIdentityAttribute";
+import { OwnRelationshipAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/OwnRelationshipAttribute";
+import { PeerIdentityAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/PeerIdentityAttribute";
+import { PeerRelationshipAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/PeerRelationshipAttribute";
+import { ThirdPartyRelationshipAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/ThirdPartyRelationshipAttribute";
 import {
     AttributeAlreadySharedAcceptResponseItemJSON,
     AttributeSuccessionAcceptResponseItemJSON,
