+++ packages/runtime/src/useCases/consumption/attributes/ChangeDefaultOwnIdentityAttribute.ts
@@ -1,5 +1,7 @@
 import { Result } from "@js-soft/ts-utils";
-import { AttributesController, LocalAttribute, OwnIdentityAttribute } from "@nmshd/consumption";
+import { AttributesController } from "@nmshd/consumption";
+import { LocalAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/LocalAttribute";
+import { OwnIdentityAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/OwnIdentityAttribute";
 import { CoreId } from "@nmshd/core-types";
 import { LocalAttributeDTO } from "@nmshd/runtime-types";
 import { AccountController } from "@nmshd/transport";
