+++ packages/runtime/src/useCases/consumption/attributes/DeleteSharedAttributesForRejectedOrRevokedRelationship.ts
@@ -1,5 +1,6 @@
 import { Result } from "@js-soft/ts-utils";
-import { AttributesController, ForwardableAttribute } from "@nmshd/consumption";
+import { AttributesController } from "@nmshd/consumption";
+import { ForwardableAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/ForwardableAttribute";
 import { CoreId } from "@nmshd/core-types";
 import { AccountController, Relationship, RelationshipsController, RelationshipStatus } from "@nmshd/transport";
 import { Inject } from "@nmshd/typescript-ioc";
