+++ packages/runtime/src/useCases/consumption/attributes/GetAttribute.ts
@@ -1,5 +1,6 @@
 import { Result } from "@js-soft/ts-utils";
-import { AttributesController, LocalAttribute } from "@nmshd/consumption";
+import { AttributesController } from "@nmshd/consumption";
+import { LocalAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/LocalAttribute";
 import { CoreId } from "@nmshd/core-types";
 import { LocalAttributeDTO } from "@nmshd/runtime-types";
 import { Inject } from "@nmshd/typescript-ioc";
