+++ packages/runtime/src/useCases/consumption/attributes/NotifyPeerAboutOwnIdentityAttributeSuccession.ts
@@ -1,5 +1,7 @@
 import { Result } from "@js-soft/ts-utils";
-import { AttributesController, ConsumptionIds, LocalAttribute, OwnIdentityAttribute } from "@nmshd/consumption";
+import { AttributesController, ConsumptionIds } from "@nmshd/consumption";
+import { LocalAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/LocalAttribute";
+import { OwnIdentityAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/OwnIdentityAttribute";
 import { Notification, PeerAttributeSucceededNotificationItem } from "@nmshd/content";
 import { CoreAddress, CoreId } from "@nmshd/core-types";
 import { LocalAttributeDTO } from "@nmshd/runtime-types";
