+++ packages/runtime/src/useCases/consumption/attributes/SucceedOwnIdentityAttribute.ts
@@ -1,5 +1,6 @@
 import { Result } from "@js-soft/ts-utils";
-import { AttributesController, ConsumptionCoreErrors, OwnIdentityAttribute, OwnIdentityAttributeSuccessorParamsJSON } from "@nmshd/consumption";
+import { AttributesController, ConsumptionCoreErrors, OwnIdentityAttributeSuccessorParamsJSON } from "@nmshd/consumption";
+import { OwnIdentityAttribute } from "@nmshd/consumption/src/modules/attributes/local/attributeTypes/OwnIdentityAttribute";
 import { AttributeValues } from "@nmshd/content";
 import { CoreId } from "@nmshd/core-types";
 import { LocalAttributeDTO } from "@nmshd/runtime-types";
